Refactor javascript.
This loads the javascript externally rather
than inline, and also gets rid of templating
in the javascript.  All info that the javascript
needs is retrieved via 'data-*' attributes on the
div.

This also uses 'jekyll-mapping' as a class instead of
an id, so that the plugin can handle multiple maps on
the same page (on index view, for example).

Known issues: it uses document.getElementsByClassName
which prob doesn't work in all IE versions.  Doesn't yet
support 'locations' but does support 'layers' and 'lat/long'.